requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
flower==2.0.1
humanize==4.11.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
